rework namespaces to follow microsoft guidelines
As mentioned here https://learn.microsoft.com/en-us/dotnet/standard/design-guidelines/names-of-namespaces
the namespaces must be prefixed by a company or an identifier that avoid namespace collisions.
For this reason, the repo was updated to prefix the namespaces with Glihm, and also the internal projects
like Samples and TestApp were renamed for more readability.